Update imports after Cypher refactoring.
