+++ cypher/cypher-docs/src/test/scala/org/neo4j/cypher/docgen/tooling/contentAndResultMerger.scala
@@ -19,8 +19,8 @@
  */
 package org.neo4j.cypher.docgen.tooling
 
-import org.neo4j.cypher.internal.v4_0.util.Rewritable._
-import org.neo4j.cypher.internal.v4_0.util.{Rewriter, bottomUp}
+import org.neo4j.cypher.internal.util.Rewritable._
+import org.neo4j.cypher.internal.util.{Rewriter, bottomUp}
 
 /**
  * Takes the document tree and the execution results and rewrites the
